+++ lib/features/game/domain/repositories/board_repository.dart
@@ -1,8 +1,9 @@
 import '../../../../core/enums/direction.dart';
 import '../entities/board.dart';
 
-abstract class BoardRepository{
+abstract class BoardRepository {
   Future<Board> getCurrentBoard();
   Future<void> resetBoard();
   Future<Board> updateBoard(Board board, Direction direction);
+  Future<int> getHighscore();
 }