+++ lib/features/game/domain/usecases/get_highscore.dart
@@ -1,0 +1,15 @@
+import 'package:meta/meta.dart';
+
+import '../../../../core/usecases/usecase.dart';
+import '../repositories/board_repository.dart';
+
+class GetHighscore implements UseCase<int, NoParams> {
+  final BoardRepository boardRepository;
+
+  GetHighscore({@required this.boardRepository})
+      : assert(boardRepository != null);
+  @override
+  Future<int> call(NoParams params) async {
+    return await boardRepository.getHighscore();
+  }
+}