+++ lib/main.dart
@@ -1,15 +1,30 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:hive/hive.dart';
+import 'package:path_provider/path_provider.dart' as path_provider;
 
 import 'core/router/route_paths.dart';
 import 'core/router/router.dart';
 import 'core/theme/custom_theme.dart';
 import 'injection_container.dart' as di;
 
-void main() {
-  di.init();
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  // no need to init hive for browser
+  if (!kIsWeb) {
+    // init Hive for mobile
+    Hive.init((await path_provider.getApplicationDocumentsDirectory()).path);
+  }
+
+  // dependency injection
+  await di.init();
+
+  // init logging of BLoC transitions
   initBlocLogging();
+
+  // run the app
   runApp(MyApp());
 }